refactor: fixed the two issues that let project down.
problem one inside create_group:

The issue is that the form is being submitted
 to the group_detail URL before the group is created,
which means the slug is not yet available.
I have changed the form action to submit
to the create_group view as it should have been.

problem two with send_email view:

The error TypeError: 'QueryDict' object is not callable and POST error occurs because you are trying to call request.POST as if it were a function. Instead, I should have accessed it like a dictionary, fixed.